STYLE: Fix warning, lost precision in cast
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=17646
git-svn-id: http://svn.slicer.org/Slicer4/trunk@17646 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Libs/MRML/vtkEventBroker.cxx
@@ -527,13 +527,13 @@ vtkCollection *vtkEventBroker::GetObservationsForObserver ( vtkObject *observer 
 //----------------------------------------------------------------------------
 int vtkEventBroker::GetNumberOfObservations ( )
 {
-  int count = 0;
+  size_t count = 0;
   ObjectToObservationVectorMap::iterator iter; 
   for(iter=this->SubjectMap.begin(); iter != this->SubjectMap.end(); iter++)  
     {
     count += iter->second.size();
     }
-  return count;
+  return static_cast<int>(count);
 }
 
 //----------------------------------------------------------------------------
@@ -544,7 +544,7 @@ vtkObservation *vtkEventBroker::GetNthObservation ( int n )
     return NULL;
     }
 
-  int count = 0;
+  size_t count = 0;
   ObjectToObservationVectorMap::iterator iter; 
   for(iter=this->SubjectMap.begin(); iter != this->SubjectMap.end(); iter++)  
     {
@@ -812,7 +812,7 @@ void vtkEventBroker::QueueObservation ( vtkObservation *observation, void *callD
 //----------------------------------------------------------------------------
 int vtkEventBroker::GetNumberOfQueuedObservations ()
 {
-  return( this->EventQueue.size() );
+  return static_cast<int>( this->EventQueue.size() );
 }
 
 //----------------------------------------------------------------------------
